requirements.txt
openai>=1.12.0
anthropic>=0.25.0
pyyaml>=6.0
python-dotenv>=1.0.0
jsonschema>=4.20.0
tqdm>=4.66.0
pandas>=2.1.0
numpy>=1.25.0
